new files

+++ src/main/java/com/atripathi/journalApp/service/JournalEntryService.java
@@ -28,8 +28,7 @@ public class JournalEntryService {
             journalEntry.setDate(LocalDateTime.now());
             JournalEntry saved = journalEntryRepository.save(journalEntry);
             user.getJournalEntries().add(saved);
-            user.setUserName(null);
-            userService.saveEntry(user);
+            userService.saveNewUser(user);
         } catch (Exception e) {
             System.out.println(e);
             throw new RuntimeException("An error occurred while saving the entry.", e);
@@ -49,11 +48,21 @@ public class JournalEntryService {
         return journalEntryRepository.findById(id);
     }
 
-    public void deleteById(ObjectId id, String userName) {
-        User user = userService.findByUserName(userName);
-        user.getJournalEntries().removeIf(x -> x.getId().equals(id));
-        userService.saveEntry(user); // user saved
-        journalEntryRepository.deleteById(id); // deleted
+    @Transactional
+    public boolean deleteById(ObjectId id, String userName) {
+        boolean removed = false;
+        try {
+            User user = userService.findByUserName(userName);
+            removed = user.getJournalEntries().removeIf(x -> x.getId().equals(id));
+            if (removed) {
+                userService.saveUser(user); // user saved
+                journalEntryRepository.deleteById(id); // deleted
+            }
+        } catch (Exception e) {
+            System.out.println(e);
+            throw new RuntimeException("An error occurred while deleting the entry. ", e);
+        }
+        return removed;
     }
 }
